Added additional calibration to Simple Menu
Added PH Expansion and Water Level expansion calibration to the Simple Menu.  The entries are:

PH Exp Calibration
Water Calibration

They are truncated due to the string constraints of the controller and display.  Also, if all addon's are enabled in the Simple Menu, they will not all show.  Currently, only 9 entries can be displayed on the screen at a time.  If all are enabled, the Setup Date/Time and Version menus will not be displayed.
This has to be fixed and improved by the menu scrolling up and down appropriately.