Added additional calibration to Standard Menu
Added PH Expansion and Water Level calibration to Standard Menu.  They are located under the Setup menu and using the following entries:

Calibrate PH Exp
Calibrate Water

The labels are truncated due to the string constraints and screen constraints of the controller.  The menus will display properly in the Setup menu because the entries were less than 8 before adding them.