Adding Threshold concept to the DCPump class and associated files
Added a PumpThreshold to Globals that can be used anywhere to limit the
value that your pump can be set to.  It can wrap a Wave profile to limit
the set value to values which are 0 or threshold to Max and can be used
in the .ino file or in calls to DCPump class.

It's intended to let users include a lower limit of 0 into the wave
profiles like NutrientTransport, for instance, and be confident that
their Tunze or Jebao pump, for instance, will not be driven at 10% of
it's max, which is an unsupported speed, by wrapping the call to
Nutrient Transport with a PumpThreshold(WaveProfile,Threshold) call
which would round down or up to avoid the "danger" zone.  All of the
function calls in ReefAngel.cpp that called wave profiles were wrapped
in this PumpThreshold function.