PWM changes, adding 16ch
Changing PWM resolution to 12-bit where supported by the chip and adding
16ch dimming module in to the code in PWM classes and also in DCPump.

Also adding highATO as a channel in DCPump and adding Else and Sine as
DCPump modes.

Lastly, changing around the values reported by the wifi because the
portal and apps and things expect them to be percents from 0-100 rather
than ints from 0-4095 which they are now with the 12-bit changes.