fix to the Wifi XML output
Previously the Wifi was outputting bytes on the PWM channels as DEC
format which has nothing past the decimal.  Now I'm dividing by 40.95
and getting floats, so we need to explicitly change the print statements
to have 0 digits past the decimal point rather than using the generic
DEC format.

Also added the 16ch expansion to the XML and JSON output.