Different default thresholds for RF Expansion users
If someone is using the RF Expansion we'll assume that they are using a
vortech which can be lowered in speed down to 0 rather than a Tunze or
Jebao which stop working reliably at a higher threshold.  If someone
sets the threshold explicitly or is using internal memory and set it in
memory that value will get used in DCPump calls.