reverted GetChannel type functions and added Raw
Made different version called GetChannelRaw that gives the full number
of bits and is used internally and set the GetChannel functions back to
their original way of doing it with 100 percent as the max value.  This
will let people continue to use these functions for things like custom
home screens.

Also fixed a bug or two in the PWM code.