Changes to DCPump class and methods
1) Added HighATO as an option for all DCPump modes
2) Added initialization for DCPump memory location of the Threshold
concept so if you're using internal memory with the RFExpansion module
enabled you will get a 0 written for threshold for the first time and if
you don't have an RFExpansion module enabled you'll get a 30 percent
written to the memory location.  This only happens the first time you
load the code that contains this new memory location (which is /mb364,
by the way).
3) Added defines for Else and Sine mode as DCPump modes and defined them
as mode 12 and 13, respectively, which matches what was added in U-App
as Else mode.
4) Outputting a new tag in the wifi output of <DCT> which reports the
threshold.  It comes in the XML and JSON right after <DCD> which is the
duration.

Tested with:

ReefAngel.DCPump.UseMemory = true;
ReefAngel.DCPump.DaylightChannel = Sync;
ReefAngel.DCPump.ActinicChannel = AntiSync;
ReefAngel.DCPump.LowATOChannel = Sync;
ReefAngel.DCPump.HighATOChannel = AntiSync;