Making LED profiles return old values
Overloaded the PWMSlope, PWMParabola, and PWMSmoothRamp functions so
that if you have an extra argument of a boolean (no matter what you put
in there) you will get an integer from 0 to 4095 returned, but if you
don't have that extra argument you get the former expected behavior of
percents from 0 to 100.

Internally, I am having the functions in RA_PWM use the higher
resolution versions by supplying that extra argument, so you don't need
to supply it on the frontend in order to get the higher resolutions from
the utility methods in RA_PWM like DaylightPWMSlope(),
Channel0PWMSlope(), etc., but if you are using the PWMSlope to do
something else, for instance cloud effects, the function will look the
same to you.