Adding gyre mode
Gradually slopes up sync pumps from Min speed to Max speed while keeping
anti-sync off, then shuts off sync and gradually slopes anti-sync up
while keeping other one off.

Note:  Duration is in minutes, not seconds, because if people are doing
a gyre they will probably want it to run for a while.  It's an int, so
technically it could be lots of seconds, but DCPump duration is a byte,
so if someone is using DCPump they could never get over 255 seconds if
it was in seconds, not minutes.