Fixed custom variable display ghosting
on the 2014 screen, when using custom variables and you change the value from a larger value (say 123) to a smaller value (say 68) you end up with a displayed value of 683. this is because the space that the value is displayed is not cleared before updating.
this fix clears the display first and then updates the value.
closes #216