Offsets were not being added correctly
If we were close to midnight and adding an offset that would cross midnight, we'd would not calculate the time properly. The same thing would occur if we start close to midnight and had an offset that pushed the time to before midnight, we would have a problem.

As part of this fix, I centralized the calculation of the onTime and offTime and the reading from memory locations. This is now done only once in the code now and much easier to read (390 lines removed).

We also had a bug in StandardActinic() which was calling SetDaylightRaw by mistake.